Refactor implementation-detail tests into backend/internal suites
